+++ Brocfile.js
@@ -24,6 +24,8 @@ app.hinting = false;
 // Use this to add additional libraries to the generated output files.
 app.import('vendor/ember-data/ember-data.js');
 
+app.import('vendor/ember-validations/index.js');
+
 // If the library that you are including contains AMD or ES6 modules that
 // you would like to import into your application please specify an
 // object with the list of modules as keys along with the exports of each